+++ crates/contracts/src/finance/dcd/build_witness.rs
