+++ crates/contracts/src/finance/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "finance-dcd")]
+pub mod dcd;
+#[cfg(feature = "finance-option-offer")]
+pub mod option_offer;
+#[cfg(feature = "finance-options")]
+pub mod options;
