+++ crates/contracts/src/finance/option_offer/build_witness.rs
