+++ crates/contracts/src/finance/options/build_witness.rs
