+++ crates/contracts/src/sdk/finance/mod.rs
@@ -1,0 +1,9 @@
+#[cfg(feature = "finance-option-offer")]
+mod option_offer;
+#[cfg(feature = "finance-options")]
+mod options;
+
+#[cfg(feature = "finance-option-offer")]
+pub use option_offer::*;
+#[cfg(feature = "finance-options")]
+pub use options::*;
