+++ crates/contracts/src/sdk/finance/options/creation_option.rs
@@ -1,5 +1,5 @@
-use crate::options::OptionsArguments;
-use crate::options::get_options_address;
+use crate::finance::options::OptionsArguments;
+use crate::finance::options::get_options_address;
 
 use crate::error::TransactionBuildError;
