+++ crates/contracts/src/sdk/finance/options/settlement_option.rs
@@ -1,5 +1,5 @@
-use crate::options::OptionsArguments;
-use crate::options::build_witness::OptionBranch;
+use crate::finance::options::OptionsArguments;
+use crate::finance::options::build_witness::OptionBranch;
 
 use crate::error::TransactionBuildError;
