+++ crates/contracts/src/sdk/mod.rs
@@ -1,16 +1,12 @@
 #[cfg(feature = "sdk-basic")]
 mod basic;
-#[cfg(feature = "option-offer")]
-mod option_offer;
-#[cfg(feature = "options")]
-mod options;
+#[cfg(any(feature = "finance-option-offer", feature = "finance-options"))]
+mod finance;
 
 pub mod taproot_pubkey_gen;
 pub mod validation;
 
 #[cfg(feature = "sdk-basic")]
 pub use basic::*;
-#[cfg(feature = "option-offer")]
-pub use option_offer::*;
-#[cfg(feature = "options")]
-pub use options::*;
+#[cfg(any(feature = "finance-option-offer", feature = "finance-options"))]
+pub use finance::*;
